+++ src/apiserver/client.go
@@ -3,11 +3,14 @@ package apiserver
 import (
 	"github.com/newrelic/nri-kubernetes/src/client"
 	"github.com/pkg/errors"
+	v1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/version"
 )
 
 // Client an interface for querying the k8s API server
 type Client interface {
 	GetNodeInfo(nodeName string) (*NodeInfo, error)
+	GetServerVersion() (*version.Info, error)
 }
 
 // NewClient creates a new API Server client
@@ -21,6 +24,11 @@ type clientImpl struct {
 	k8sClient client.Kubernetes
 }
 
+// GetServerVersion returns the kubernetes server version.
+func (c clientImpl) GetServerVersion() (*version.Info, error) {
+	return c.k8sClient.ServerVersion()
+}
+
 // GetNodeInfo queries the API server for information about the given node
 func (c clientImpl) GetNodeInfo(nodeName string) (*NodeInfo, error) {
 
@@ -31,17 +39,23 @@ func (c clientImpl) GetNodeInfo(nodeName string) (*NodeInfo, error) {
 	}
 
 	return &NodeInfo{
-		NodeName: node.ObjectMeta.Name,
-		Labels:   node.Labels,
+		NodeName:    node.ObjectMeta.Name,
+		Labels:      node.Labels,
+		Allocatable: node.Status.Allocatable,
+		Capacity:    node.Status.Capacity,
 	}, nil
 }
 
 // NodeInfo contains information about a specific node
 type NodeInfo struct {
-	NodeName string
-	Labels   map[string]string
+	NodeName    string
+	Labels      map[string]string
+	Allocatable v1.ResourceList
+	Capacity    v1.ResourceList
 }
 
+// IsMasterNode returns true if the NodeInfo contains the labels that
+// identify a node as master.
 func (i *NodeInfo) IsMasterNode() bool {
 	if val, ok := i.Labels["kubernetes.io/role"]; ok && val == "master" {
 		return true