+++ src/apiserver/client_test.go
@@ -9,6 +9,8 @@ import (
 	"github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	v1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/resource"
 )
 
 func getTempDir(t *testing.T) (string, func()) {
@@ -34,6 +36,18 @@ func TestFileCacheReadMiss(t *testing.T) {
 			"kubernetes.io/hostname": "MyNode",
 			"kubernetes.io/os":       "linux",
 		},
+		Allocatable: v1.ResourceList{
+			v1.ResourceCPU:              *resource.NewQuantity(2, resource.DecimalSI),
+			v1.ResourcePods:             *resource.NewQuantity(110, resource.DecimalSI),
+			v1.ResourceEphemeralStorage: *resource.NewQuantity(18211580000, resource.BinarySI),
+			v1.ResourceMemory:           *resource.NewQuantity(2033280000, resource.BinarySI),
+		},
+		Capacity: v1.ResourceList{
+			v1.ResourceCPU:              *resource.NewQuantity(2, resource.DecimalSI),
+			v1.ResourcePods:             *resource.NewQuantity(110, resource.DecimalSI),
+			v1.ResourceEphemeralStorage: *resource.NewQuantity(18211586048, resource.BinarySI),
+			v1.ResourceMemory:           *resource.NewQuantity(2033283072, resource.BinarySI),
+		},
 	}
 	client := TestAPIServer{Mem: map[string]*NodeInfo{"MyNode": myNode}}
 
@@ -51,11 +65,29 @@ func TestFileCacheReadCacheAndExpiry(t *testing.T) {
 	dir, cleanup := getTempDir(t)
 	defer cleanup()
 
+	// The resource.Quantity struct has an attribute called `s` that caches
+	// the string representation. This attribute is calculated and stored
+	// when calling the String and UnmarshalJSON functions. We need to call
+	// the String function when creating the Quantities to make the structs
+	// match and not fail on the assert.Equal function.
+	cpu := resource.NewQuantity(2, resource.DecimalSI)
+	_ = cpu.String()
+	newCPU := resource.NewQuantity(4, resource.DecimalSI)
+	_ = newCPU.String()
+	memory := resource.NewQuantity(2033283072, resource.BinarySI)
+	_ = memory.String()
+
 	myNode := &NodeInfo{
 		NodeName: "MyNode",
 		Labels: map[string]string{
 			"kubernetes.io/hostname": "MyNode",
 			"kubernetes.io/os":       "linux",
+		},
+		Allocatable: v1.ResourceList{
+			v1.ResourceCPU: *cpu,
+		},
+		Capacity: v1.ResourceList{
+			v1.ResourceCPU: *cpu,
 		},
 	}
 	myUpdatedNode := &NodeInfo{
@@ -64,6 +96,14 @@ func TestFileCacheReadCacheAndExpiry(t *testing.T) {
 			"kubernetes.io/hostname": "updatedHostname",
 			"kubernetes.io/os":       "linux",
 			"kubernetes.io/updated":  "true",
+		},
+		Allocatable: v1.ResourceList{
+			v1.ResourceCPU:    *cpu,
+			v1.ResourceMemory: *memory,
+		},
+		Capacity: v1.ResourceList{
+			v1.ResourceCPU:    *newCPU,
+			v1.ResourceMemory: *memory,
 		},
 	}
 
@@ -83,12 +123,13 @@ func TestFileCacheReadCacheAndExpiry(t *testing.T) {
 	// Reading from the cacheWrapper should still return the old version
 	node, err := cacheWrapper.GetNodeInfo("MyNode")
 	assert.NoError(t, err)
-	assert.Equal(t, node, myNode)
+
+	assert.Equal(t, myNode, node)
 
 	// While reading from the client should return the new version
 	node, err = client.GetNodeInfo("MyNode")
 	assert.NoError(t, err)
-	assert.Equal(t, node, myUpdatedNode)
+	assert.Equal(t, myUpdatedNode, node)
 
 	// The cache should reset after 1 hour, and the cacheWrapper should return the updated object
 	timeProvider.time = time.Now().Add(time.Hour * 2)