+++ src/metric/populate.go
@@ -8,6 +8,7 @@ import (
 	"github.com/newrelic/infra-integrations-sdk/sdk"
 	"github.com/newrelic/nri-kubernetes/src/data"
 	"github.com/newrelic/nri-kubernetes/src/definition"
+	"k8s.io/apimachinery/pkg/version"
 )
 
 type k8sPopulator struct {
@@ -31,8 +32,14 @@ func (e MultipleErrs) Error() string {
 }
 
 // Populate populates k8s raw data to sdk metrics.
-func (p *k8sPopulator) Populate(groups definition.RawGroups, specGroups definition.SpecGroups, i *sdk.IntegrationProtocol2, clusterName string) data.PopulateResult {
-	populatorFunc := definition.IntegrationProtocol2PopulateFunc(i, clusterName, K8sMetricSetTypeGuesser, K8sEntityMetricsManipulator, K8sClusterMetricsManipulator)
+func (p *k8sPopulator) Populate(
+	groups definition.RawGroups,
+	specGroups definition.SpecGroups,
+	i *sdk.IntegrationProtocol2,
+	clusterName string,
+	k8sVersion *version.Info,
+) data.PopulateResult {
+	populatorFunc := definition.IntegrationProtocol2PopulateFunc(i, clusterName, k8sVersion, K8sMetricSetTypeGuesser, K8sEntityMetricsManipulator, K8sClusterMetricsManipulator)
 	ok, errs := populatorFunc(groups, specGroups)
 
 	if len(errs) > 0 {